Move password reset component and page
