+++ app/(auth)/password/reset/components/PasswordReset.tsx
@@ -2,41 +2,36 @@
 import { useState } from "react";
 import { useTranslation } from "@i18n";
 import { useRouter } from "next/navigation";
-
 import { PasswordComplexitySettings } from "@zitadel/proto/zitadel/settings/v2/password_settings_pb";
 import { create } from "@zitadel/client";
 import { ChecksSchema } from "@zitadel/proto/zitadel/session/v2/session_service_pb";
-import { changePassword, sendPassword } from "@lib/server/password";
 
-import { Alert, ErrorStatus } from "@clientComponents/forms";
 import { PasswordValidationForm } from "@components/PasswordValidation/PasswordValidationForm";
+import { Alert, ErrorStatus } from "@clientComponents/forms";
+import { changePassword, sendPassword } from "@lib/server/password";
 
 export function PasswordReset({
   userId,
-  code,
   passwordComplexitySettings,
   organization,
-  requestId,
   loginName,
 }: {
   userId?: string;
-  code?: string;
   passwordComplexitySettings?: PasswordComplexitySettings;
   organization?: string;
-  requestId?: string;
   loginName?: string;
 }) {
   const { t } = useTranslation(["password"]);
   const router = useRouter();
   const [error, setError] = useState("");
 
-  const successCallback = async ({ password }: { password: string }) => {
+  const submitPasswordForm = async ({ password, code }: { password: string; code?: string }) => {
     if (!userId) return;
 
     const payload: { userId: string; password: string; code?: string } = {
       userId: userId,
       password,
-      code,
+      ...(code ? { code } : {}),
     };
 
     const changeResponse = await changePassword(payload).catch(() =>
@@ -53,28 +48,12 @@ export function PasswordReset({
       return;
     }
 
-    ContinueAuthenticationFlow({ password });
-  };
-
-  const ContinueAuthenticationFlow = async ({ password }: { password: string }) => {
-    const params = new URLSearchParams({});
-
-    if (loginName) {
-      params.append("loginName", loginName);
-    }
-    if (organization) {
-      params.append("organization", organization);
-    }
-
-    await new Promise((resolve) => setTimeout(resolve, 2000)); // Wait for a second to avoid eventual consistency issues with an initial password being set
-
     const passwordResponse = await sendPassword({
       loginName: loginName ?? "",
       organization,
       checks: create(ChecksSchema, {
         password: { password },
       }),
-      requestId,
     }).catch(() => setError(t("reset.errors.couldNotVerifyPassword")));
 
     if (passwordResponse && "error" in passwordResponse && passwordResponse.error) {
@@ -96,7 +75,8 @@ export function PasswordReset({
       {error && <Alert type={ErrorStatus.ERROR}>{error}</Alert>}
       <PasswordValidationForm
         passwordComplexitySettings={passwordComplexitySettings}
-        successCallback={successCallback}
+        successCallback={submitPasswordForm}
+        requireConfirmationCode={true}
       />
     </>
   );
