+++ app/(auth)/password/reset/page.tsx
@@ -1,53 +1,55 @@
+import { getServiceUrlFromHeaders } from "@lib/service-url";
+import { loadMostRecentSession } from "@lib/session";
+import { getPasswordComplexitySettings } from "@lib/zitadel";
+import { Session } from "@zitadel/proto/zitadel/session/v2/session_pb";
 import { Metadata } from "next";
 import { serverTranslation } from "@i18n/server";
 import { headers } from "next/headers";
-
-import { getPasswordComplexitySettings } from "@lib/zitadel";
-import { getServiceUrlFromHeaders } from "@lib/service-url";
 import { getSessionCredentials } from "@lib/cookies";
-import { loadSessionById } from "@lib/session";
 
 import { AuthPanel } from "@serverComponents/globals/AuthPanel";
 import { PasswordReset } from "./components/PasswordReset";
 
 export async function generateMetadata(): Promise<Metadata> {
-  const { t } = await serverTranslation("password.set");
-  return { title: t("title") };
+  const { t } = await serverTranslation("password");
+  return { title: t("reset.title") };
 }
 
-export default async function Page({
-  searchParams,
-}: {
-  searchParams: Promise<Record<string | number | symbol, string | undefined>>;
-}) {
+export default async function Page() {
+  const { loginName, organization } = await getSessionCredentials();
   const _headers = await headers();
   const { serviceUrl } = getServiceUrlFromHeaders(_headers);
 
-  const params = await searchParams;
-  const code = params.code;
-
-  const { sessionId, loginName, organization } = await getSessionCredentials();
-  const sessionFactors = await loadSessionById(serviceUrl, sessionId, organization);
+  // also allow no session to be found (ignoreUnkownUsername)
+  let session: Session | undefined;
+  if (loginName) {
+    session = await loadMostRecentSession({
+      serviceUrl,
+      sessionParams: {
+        loginName,
+        organization,
+      },
+    });
+  }
 
   const passwordComplexitySettings = await getPasswordComplexitySettings({
     serviceUrl,
     organization,
   });
 
-  if (!loginName || !sessionId || !organization) {
+  if (!loginName || !session || !organization) {
     throw new Error("No session.");
   }
 
   return (
     <AuthPanel
-      titleI18nKey="password.title"
-      descriptionI18nKey="password.description"
-      namespace="reset"
+      titleI18nKey="reset.title"
+      descriptionI18nKey="reset.description"
+      namespace="password"
     >
       <PasswordReset
-        userId={sessionFactors.factors?.user?.id}
+        userId={session.factors?.user?.id}
         passwordComplexitySettings={passwordComplexitySettings}
-        code={code}
         organization={organization}
         loginName={loginName}
       />
